@@ -20,7 +20,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=2.18.2 <3.0.0'
+  sdk: '^3.5.4'
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -32,29 +32,29 @@ dependencies:
   flutter:
     sdk: flutter
   cupertino_icons: ^1.0.2
-  firebase_core: ^2.2.0
-  firebase_database: ^10.2.1
-  firebase_auth: ^4.1.0
-  google_sign_in: ^5.4.2
-  flutter_facebook_auth: 4.4.1
-  cloud_firestore: ^4.1.0
-  http: ^0.13.5
+  firebase_core: ^3.12.1
+  firebase_database: ^11.3.4
+  firebase_auth: ^5.5.1
+  google_sign_in: ^6.2.2
+  flutter_facebook_auth: ^7.1.1
+  cloud_firestore: ^5.6.5
+  http: ^1.3.0
   animations: ^2.0.7
   carousel_slider: ^5.0.0
-  image_cropper: ^3.0.1
-  image_picker: ^0.8.6+1
+  image_cropper: ^9.0.0
+  image_picker: ^1.1.2
   shared_preferences: ^2.0.17
-  firebase_storage: ^11.0.0
+  firebase_storage: ^12.4.4
   flutter_cache_manager: ^3.3.0
   path_provider: ^2.0.14
-  audioplayers: ^4.0.1
-  flutter_secure_storage: ^8.0.0
-  purchases_flutter: ^5.0.0
+  audioplayers: ^6.4.0
+  flutter_secure_storage: ^9.2.4
+  purchases_flutter: ^8.6.1
   flutter_riverpod: ^2.3.6
   video_player: ^2.7.0
-  algolia_helper_flutter: ^0.3.3
+  algolia_helper_flutter: ^1.2.0
   infinite_scroll_pagination: ^5.0.0
-  google_mobile_ads: ^3.0.0
+  google_mobile_ads: ^5.3.1
   ffi: ^2.1.3
   win32: ^5.10.0
 
@@ -69,7 +69,7 @@ dev_dependencies:
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  flutter_lints: ^2.0.0
+  flutter_lints: ^5.0.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
