@@ -43,5 +43,5 @@ String CF_SUBCOLLECTION_FRIENDS = 'friends';
 String CF_DOC_PENDING_FRIEND_REQUESTS = 'pendingFriendRequests';
 
 // End point
-//String END_POINT = "https://brick-hold-em-api.onrender.com";
-String END_POINT = "http://192.168.1.14:3000"; // for when local
+// String END_POINT = "https://brick-hold-em-api.onrender.com";
+String END_POINT = "http://192.168.1.18:3000"; // for when local
